+++ paddle/phi/kernels/cpu/uniform_inplace_kernel.cc
@@ -21,8 +21,8 @@ namespace phi {
 template <typename T, typename Context>
 void UniformInplaceKernel(const Context& dev_ctx,
                           const DenseTensor& x UNUSED,
-                          float min,
-                          float max,
+                          double min,
+                          double max,
                           int seed,
                           int diag_num UNUSED,
                           int diag_step UNUSED,
