+++ paddle/phi/kernels/gaussian_kernel.h
@@ -32,8 +32,8 @@ PADDLE_API void GaussianKernel(const Context& dev_ctx,
 template <typename T, typename Context>
 void GaussianInplaceKernel(const Context& dev_ctx,
                            const DenseTensor& x,
-                           float mean,
-                           float std,
+                           double mean,
+                           double std,
                            int seed,
                            DenseTensor* out);
