+++ paddle/phi/kernels/gpu/gaussian_inplace_grad_kernel.cu
@@ -23,8 +23,8 @@ namespace phi {
 template <typename T, typename Context>
 void GaussianInplaceGradKernel(const Context& dev_ctx,
                                const DenseTensor& out_grad,
-                               float mean,
-                               float std,
+                               double mean,
+                               double std,
                                int seed,
                                DenseTensor* x_grad) {
   float value = static_cast<float>(0.0f);
