+++ paddle/phi/kernels/xpu/gaussian_kernel.cc
@@ -52,8 +52,8 @@ void GaussianKernel(const Context& dev_ctx,
 template <typename T, typename Context>
 void GaussianInplaceKernel(const Context& dev_ctx,
                            const DenseTensor& x,
-                           float mean,
-                           float std,
+                           double mean,
+                           double std,
                            int seed,
                            DenseTensor* out) {
   T* data = dev_ctx.template Alloc<T>(out);
@@ -67,8 +67,8 @@ void GaussianInplaceKernel(const Context& dev_ctx,
 
   int r = xpu::normal_<XPUType>(dev_ctx.x_context(),
                                 reinterpret_cast<XPUType*>(data),
-                                mean,
-                                std,
+                                static_cast<float>(mean),
+                                static_cast<float>(std),
                                 out->numel(),
                                 real_seed);
   PADDLE_ENFORCE_XDNN_SUCCESS(r, "normal");
