+++ paddle/phi/kernels/xpu/uniform_inplace_grad_kernel.cc
@@ -20,8 +20,8 @@ namespace phi {
 template <typename T, typename Context>
 void XPUUniformRandomInplaceGradKernel(const Context& dev_ctx,
                                        const DenseTensor& out_grad,
-                                       float min UNUSED,
-                                       float max UNUSED,
+                                       double min UNUSED,
+                                       double max UNUSED,
                                        int seed UNUSED,
                                        int diag_num UNUSED,
                                        int diag_step UNUSED,
